+++ linopy/variables.py
@@ -24,6 +24,7 @@ from linopy.common import (
     _merge_inplace,
     forward_as_properties,
     has_optimized_model,
+    head_tail_range,
     is_constant,
     print_coord,
     print_single_variable,
@@ -117,6 +118,8 @@ class Variable:
     __slots__ = ("_labels", "_model")
     __array_ufunc__ = None
 
+    _fill_value = -1
+
     def __init__(self, labels: DataArray, model: Any):
         """
         Initialize the Constraint.
@@ -181,52 +184,60 @@ class Variable:
         """
         Print the variable arrays.
         """
-        lower = self.lower.values
-        upper = self.upper.values
-
         # don't loop over all values if not necessary
+        # TODO: check that we can skip this
         if self.size == 1:
-            header = f"{self.type}\n{'-' * len(self.type)}"
-            lower = lower.item()
-            upper = upper.item()
-            data_string = print_single_variable(lower, upper, self.name, self.type)
+            header = f"Variable\n{'-' * len('Variable')}"
+            if self.lower.size > 1:
+                coord = {k: v for k, v in self.coords.items() if k in self.lower.dims}
+                lower = self.lower.sel(coord).item()
+            else:
+                lower = self.lower.item()
+            if self.upper.size > 1:
+                coord = {k: v for k, v in self.coords.items() if k in self.upper.dims}
+                upper = self.upper.sel(coord).item()
+            else:
+                upper = self.upper.item()
+            coord = []
+            data_string = print_single_variable(self, self.name, coord, lower, upper)
             return f"{header}\n{data_string}"
 
         # print only a few values
         max_print = 14
         split_at = max_print // 2
-        to_print = np.flatnonzero(self.mask)
-        truncate = len(to_print) > max_print
-        if truncate:
-            to_print = np.hstack([to_print[:split_at], to_print[-split_at:]])
+        to_print = head_tail_range(self.size, max_print)
 
         # create string, we use numpy to get the indexes
-        data_string = ""
         idx = np.unravel_index(to_print, self.shape)
-        indexes = np.stack(idx)
         coords = [self.indexes[self.dims[i]][idx[i]] for i in range(len(self.dims))]
+        coords = list(zip(*coords))
+        labels = np.ravel(self.labels.values)[to_print]
 
-        # loop over all values to print
-        for i in range(len(to_print)):
-            # this is the index for the labels array
-            ix = tuple(indexes[..., i])
-            # lower and upper bounds might only be defined for some dimensions
-            lix = tuple(
-                ix[i] for i in range(self.ndim) if self.dims[i] in self.lower.dims
-            )
-            uix = tuple(
-                ix[i] for i in range(self.ndim) if self.dims[i] in self.upper.dims
-            )
+        data_string = ""
+        for i, coord in enumerate(coords):
 
-            # create coordinate string
-            coord = [c[i] for c in coords]
+            label = labels[i]
             coord_string = print_coord(coord)
-            var_string = f"{self.name}{coord_string}"
-            data_string += print_single_variable(
-                lower[lix], upper[uix], var_string, self.type
-            )
 
-            if i == split_at - 1 and truncate:
+            if label != -1:
+                vname, vcoord = self.model.variables.get_label_position(label)
+
+                # get lower and upper bounds, which might have less dimensions
+                lcoord = [
+                    c for i, c in enumerate(vcoord) if self.dims[i] in self.lower.dims
+                ]
+                ucoord = [
+                    c for i, c in enumerate(vcoord) if self.dims[i] in self.upper.dims
+                ]
+                lower = self.lower.loc[tuple(lcoord)].item()
+                upper = self.upper.loc[tuple(ucoord)].item()
+                var_string = print_single_variable(self, vname, vcoord, lower, upper)
+            else:
+                var_string = "None"
+
+            data_string += f"\n{coord_string}: {var_string}"
+
+            if i == split_at - 1 and self.size > max_print:
                 data_string += "\n\t\t..."
 
         # create shape string
@@ -236,8 +247,8 @@ class Variable:
         shape_string = f"({shape_string})"
         n_masked = (~self.mask).sum().item()
         mask_string = f" - {n_masked} masked entries" if n_masked else ""
-        header = f"{self.type} {shape_string}{mask_string}\n" + "-" * (
-            len(self.type) + len(shape_string) + len(mask_string) + 1
+        header = f"Variable {shape_string}{mask_string}\n" + "-" * (
+            len("Variable") + len(shape_string) + len(mask_string) + 1
         )
         return f"{header}{data_string}"
 
@@ -452,6 +463,14 @@ class Variable:
         else:
             return "Continuous Variable"
 
+    @classmethod
+    @property
+    def fill_value(self):
+        """
+        Return the fill value of the variable.
+        """
+        return self._fill_value
+
     @property
     def mask(self):
         """
@@ -464,7 +483,7 @@ class Variable:
         -------
         xr.DataArray
         """
-        return (self.labels != -1).astype(bool)
+        return (self.labels != self._fill_value).astype(bool)
 
     @property
     def upper(self):
@@ -589,7 +608,63 @@ class Variable:
         -------
         linopy.Variable
         """
+        if isinstance(other, Variable):
+            other = other.labels
+        elif isinstance(other, ScalarVariable):
+            other = other.label
         return self.__class__(self.labels.where(cond, other, **kwargs), self.model)
+
+    def ffill(self, dim, limit=None):
+        """
+        Forward fill the variable along a dimension.
+
+        This operation call ``xarray.DataArray.ffill`` but ensures preserving
+        the linopy.Variable type.
+
+        Parameters
+        ----------
+        dim : str
+            Dimension over which to forward fill.
+        limit : int, optional
+            Maximum number of consecutive NaN values to forward fill. Must be greater than or equal to 0.
+
+        Returns
+        -------
+        linopy.Variable
+        """
+        labels = (
+            self.labels.where(self.labels != -1)
+            .ffill(dim, limit=limit)
+            .fillna(-1)
+            .astype(int)
+        )
+        return self.__class__(labels, self.model)
+
+    def bfill(self, dim, limit=None):
+        """
+        Backward fill the variable along a dimension.
+
+        This operation call ``xarray.DataArray.bfill`` but ensures preserving
+        the linopy.Variable type.
+
+        Parameters
+        ----------
+        dim : str
+            Dimension over which to backward fill.
+        limit : int, optional
+            Maximum number of consecutive NaN values to backward fill. Must be greater than or equal to 0.
+
+        Returns
+        -------
+        linopy.Variable
+        """
+        labels = (
+            self.labels.where(self.labels != -1)
+            .bfill(dim, limit=limit)
+            .fillna(-1)
+            .astype(int)
+        )
+        return self.__class__(labels, self.model)
 
     def sanitize(self):
         """
@@ -611,10 +686,6 @@ class Variable:
 
     assign_coords = varwrap(DataArray.assign_coords)
 
-    astype = varwrap(DataArray.astype)
-
-    bfill = varwrap(DataArray.bfill)
-
     broadcast_like = varwrap(DataArray.broadcast_like)
 
     compute = varwrap(DataArray.compute)
@@ -624,8 +695,6 @@ class Variable:
     drop_sel = varwrap(DataArray.drop_sel)
 
     drop_isel = varwrap(DataArray.drop_isel)
-
-    ffill = varwrap(DataArray.ffill)
 
     fillna = varwrap(DataArray.fillna)
 
@@ -971,6 +1040,8 @@ class ScalarVariable:
         self._model = model
 
     def __repr__(self) -> str:
+        if self.label == -1:
+            return "ScalarVariable: None"
         name, coord = self.model.variables.get_label_position(self.label)
         coord_string = print_coord(coord)
         return f"ScalarVariable: {name}{coord_string}"